[#5] Reorganize QoS struct conversions
