+++ rmw_iceoryx2_cxx/CMakeLists.txt
@@ -58,7 +58,6 @@ endforeach()
 
 add_library(${PROJECT_NAME} SHARED
   src/impl/common/names.cpp
-  src/impl/common/qos.cpp
   src/impl/common/qos_attributes.cpp
   src/impl/common/qos_codec.cpp
   src/impl/common/qos_diagnostics.cpp
